# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaboration Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WorkflowTemplate {
  id: number;
  name: string;
  sector: string;
  description: string;
  outputFinal: string;
  /** profesional | organisasi */
  mode: string;
  /** @nullable */
  stageCount?: number | null;
  /** @nullable */
  workroomCount?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface WorkflowTemplateInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sector: string;
  description: string;
  outputFinal: string;
  mode: string;
}

export interface WorkflowTemplateUpdate {
  name?: string;
  sector?: string;
  description?: string;
  outputFinal?: string;
  mode?: string;
}

export interface Agent {
  id: number;
  name: string;
  slugUrl: string;
  category: string;
  /** dialog | specialist | multi_agent */
  agentType: string;
  /** strategis | skeptis | eksekutor | narasumber | pack_compiler */
  functionRole: string;
  /** @nullable */
  systemPrompt?: string | null;
  active: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface AgentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slugUrl: string;
  category: string;
  agentType: string;
  functionRole: string;
  systemPrompt?: string;
  active?: boolean;
}

export interface AgentUpdate {
  name?: string;
  slugUrl?: string;
  category?: string;
  agentType?: string;
  functionRole?: string;
  systemPrompt?: string;
  active?: boolean;
}

export interface Workroom {
  id: number;
  name: string;
  templateId: number;
  templateName: string;
  sector: string;
  /** active | paused | completed | blocked */
  status: string;
  currentStageName: string;
  /** 0-100 percent */
  progress: number;
  /** @nullable */
  objective?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface WorkroomStage {
  id: number;
  workroomId: number;
  order: number;
  name: string;
  /** kerja | gate */
  stageType: string;
  /** sequential | multiclaw | openclaw */
  operandPattern: string;
  /** pending | active | awaiting_gate | approved | rejected | completed */
  status: string;
  /** @nullable */
  gateDecision?: string | null;
  /** @nullable */
  gateNote?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface WorkroomTask {
  id: number;
  workroomId: number;
  stageId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  assigneeRole?: string | null;
  /** @nullable */
  agentId?: number | null;
  /** low | medium | high */
  priority: string;
  /** todo | doing | blocked | done */
  status: string;
  /** @nullable */
  output?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface WorkroomDetail {
  id: number;
  name: string;
  templateId: number;
  templateName: string;
  sector: string;
  status: string;
  currentStageName: string;
  progress: number;
  /** @nullable */
  objective?: string | null;
  stages: WorkroomStage[];
  tasks: WorkroomTask[];
  createdAt: string;
  updatedAt: string;
}

export interface WorkroomInput {
  /** @minLength 1 */
  name: string;
  templateId: number;
  objective?: string;
}

export interface WorkroomUpdate {
  name?: string;
  status?: string;
  objective?: string;
}

export interface WorkroomStageUpdate {
  status?: string;
  gateDecision?: string;
  gateNote?: string;
}

export interface WorkroomTaskInput {
  stageId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  assigneeRole?: string;
  agentId?: number;
  priority?: string;
  status?: string;
  output?: string;
}

export interface WorkroomTaskUpdate {
  title?: string;
  description?: string;
  assigneeRole?: string;
  agentId?: number;
  priority?: string;
  status?: string;
  output?: string;
}

export interface ActivityEntry {
  id: number;
  workroomId: number;
  eventType: string;
  description: string;
  /** @nullable */
  actor?: string | null;
  createdAt: string;
}

export interface ActivityEntryInput {
  eventType: string;
  description: string;
  actor?: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface SectorCount {
  sector: string;
  count: number;
}

export interface DashboardSummary {
  totalWorkrooms: number;
  activeWorkrooms: number;
  completedWorkrooms: number;
  totalTemplates: number;
  totalAgents: number;
  workroomsByStatus: StatusCount[];
  workroomsBySector: SectorCount[];
}