# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Collaboration Builder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all workflow templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowTemplate"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a workflow template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowTemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTemplate"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a workflow template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTemplate"
        "404":
          description: Not found
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a workflow template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowTemplateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTemplate"
        "404":
          description: Not found
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a workflow template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Agents ─────────────────────────────────────────────────────────────────
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Workrooms ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    WorkflowTemplate:
      type: object
      required: [id, name, sector, description, outputFinal, mode, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        sector:
          type: string
        description:
          type: string
        outputFinal:
          type: string
        mode:
          type: string
          description: "profesional | organisasi"
        stageCount:
          type: ["integer", "null"]
        workroomCount:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    WorkflowTemplateInput:
      type: object
      required: [name, sector, description, outputFinal, mode]
      properties:
        name:
          type: string
          minLength: 1
        sector:
          type: string
          minLength: 1
        description:
          type: string
        outputFinal:
          type: string
        mode:
          type: string
    WorkflowTemplateUpdate:
      type: object
      properties:
        name:
          type: string
        sector:
          type: string
        description:
          type: string
        outputFinal:
          type: string
        mode:
          type: string

    # ── Agent ────────────────────────────────────────────────────────────────
    Agent:
      type: object
      required: [id, name, slugUrl, category, agentType, functionRole, active, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slugUrl:
          type: string
        category:
          type: string
        agentType:
          type: string
          description: "dialog | specialist | multi_agent"
        functionRole:
          type: string
          description: "strategis | skeptis | eksekutor | narasumber | pack_compiler"
        systemPrompt:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    AgentInput:
      type: object
      required: [name, slugUrl, category, agentType, functionRole]
      properties:
        name:
          type: string
          minLength: 1
        slugUrl:
          type: string
          minLength: 1
        category:
          type: string
        agentType:
          type: string
        functionRole:
          type: string
        systemPrompt:
          type: string
        active:
          type: boolean
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
        slugUrl:
          type: string
        category:
          type: string
        agentType:
          type: string
        functionRole:
          type: string
        systemPrompt:
          type: string
        active:
          type: boolean

    # ── Workroom ─────────────────────────────────────────────────────────────
...