# lib/api-zod/src/generated/types/agent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaboration Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface Agent {
  id: number;
  name: string;
  slugUrl: string;
  category: string;
  /** dialog | specialist | multi_agent */
  agentType: string;
  /** strategis | skeptis | eksekutor | narasumber | pack_compiler */
  functionRole: string;
  /** @nullable */
  systemPrompt?: string | null;
  active: boolean;
  createdAt: string;
  updatedAt: string;
}