# lib/api-zod/src/generated/types/agentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaboration Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slugUrl: string;
  category: string;
  agentType: string;
  functionRole: string;
  systemPrompt?: string;
  active?: boolean;
}