# lib/api-zod/src/generated/types/workflowTemplate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaboration Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkflowTemplate {
  id: number;
  name: string;
  sector: string;
  description: string;
  outputFinal: string;
  /** profesional | organisasi */
  mode: string;
  /** @nullable */
  stageCount?: number | null;
  /** @nullable */
  workroomCount?: number | null;
  createdAt: string;
  updatedAt: string;
}