# lib/api-zod/src/generated/types/workflowTemplateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaboration Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkflowTemplateInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sector: string;
  description: string;
  outputFinal: string;
  mode: string;
}