# lib/api-zod/src/generated/types/workflowTemplateUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaboration Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkflowTemplateUpdate {
  name?: string;
  sector?: string;
  description?: string;
  outputFinal?: string;
  mode?: string;
}